MAT-9462: Move mongo aggregation creation methods to new class to be excluded from coverage checks.
 We don't own the mongo driver and, thus, should be mocking these methods during unit tests.
