+++ .github/.codecov.yml
@@ -14,3 +14,5 @@ coverage:
             base: auto
             paths:
             - "src"
+ignore:
+  - "src/main/java/cms/gov/madie/measure/utils/SearchAggregationUtils.java"
