+++ src/main/java/cms/gov/madie/measure/repositories/MeasureSetSearchRepositoryImpl.java
@@ -2,6 +2,7 @@ package cms.gov.madie.measure.repositories;
 
 import cms.gov.madie.measure.dto.MeasureListDTO;
 import cms.gov.madie.measure.dto.MeasureSearchCriteria;
+import cms.gov.madie.measure.utils.SearchAggregationUtils;
 import cms.gov.madie.measure.utils.SearchUtils;
 import org.apache.commons.collections4.CollectionUtils;
 import org.apache.commons.lang3.StringUtils;
@@ -46,7 +47,7 @@ public class MeasureSetSearchRepositoryImpl implements MeasureSetSearchRepositor
         && StringUtils.isNotBlank(measureSearchCriteria.getSearchField())) {
       if (CollectionUtils.isEmpty(measureSearchCriteria.getOptionalSearchProperties())
           || measureSearchCriteria.getOptionalSearchProperties().contains("cmsId")) {
-        aggregationOperations.add(SearchUtils.addCmsIdDisplayField());
+        aggregationOperations.add(SearchAggregationUtils.addCmsIdDisplayField());
       }
       SearchUtils.appendAdditionalSearchCriteria(measureCriteria, measureSearchCriteria);
     }
@@ -59,7 +60,7 @@ public class MeasureSetSearchRepositoryImpl implements MeasureSetSearchRepositor
       // composite measure components search
       if (CollectionUtils.isNotEmpty(measureSearchCriteria.getAllowedScoringTypes())) {
         aggregationOperations.add(
-            SearchUtils.createScoringTypeFilter(measureSearchCriteria.getAllowedScoringTypes()));
+          SearchAggregationUtils.createScoringTypeFilter(measureSearchCriteria.getAllowedScoringTypes()));
       }
     }
